move deep_update
