+++ qcodes/data/hdf5_format.py
@@ -7,8 +7,8 @@ import h5py
 import numpy as np
 
 import qcodes as qc
-from qcodes.utils import NumpyJSONEncoder
-from qcodes.utils.helpers import deep_update
+from qcodes.utils import NumpyJSONEncoder, deep_update
+
 from .data_array import DataArray
 from .format import Formatter
